+++ src/audio/speaker.c
@@ -1272,7 +1272,7 @@ bool Spkr_DSInit()
 	HRESULT hr = DSGetSoundBuffer(&SpeakerVoice, DSBCAPS_CTRLVOLUME, g_dwDSSpkrBufferSize, SPKR_SAMPLE_RATE, 1);
 	if(FAILED(hr))
 	{
-		if(g_fh) fprintf(g_fh, "Spkr: DSGetSoundBuffer failed (%08X)\n",hr);
+		if(g_fh) fprintf(g_fh, "Spkr: DSGetSoundBuffer failed (%08X)\n",(unsigned int)hr);
 		return false;
 	}
